Remove hard-coded digest and make store a mod dir.
